requirements.txt
absl-py==2.1.0
accelerate==1.3.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.2
aiohttp==3.10.0
aiohttp-cors==0.7.0
aiosignal==1.3.1
airportsdata==20241001
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==12.3.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.3
black==25.1.0
blake3==1.0.4
blinker==1.8.2
bracex==2.4
build==1.2.1
cachetools==5.5.1
cdifflib==1.2.6
certifi==2025.1.31
cffi==1.16.0
charset-normalizer==3.4.1
clarabel==0.9.0
click==8.1.8
cloudpickle==3.1.0
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
colorful==0.5.6
comm==0.2.2
compressed-tensors==0.9.1
ConfigArgParse==1.7
contourpy==1.2.1
coverage==7.6.12
crcmod==1.7
cryptography==44.0.2
cupy-cuda12x==13.3.0
cvxpy==1.5.2
cycler==0.12.1
Cython==3.0.10
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dashscope==1.22.1
dataclasses-json==0.6.7
datasets==3.2.0
decorator==4.4.2
decord==0.6.0
deepspeed==0.15.4
depth_diff_gaussian_rasterization_min==0.0.0
depyf==0.18.0
diffusers==0.33.1
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
DistVAE==0.0.0b5
dnspython==2.6.1
docker-pycreds==0.4.0
easydict==1.13
ecos==2.0.14
editdistance==0.8.1
einops==0.8.0
einops-exts==0.0.4
einx==0.3.0
email_validator==2.2.0
encodec==0.1.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
fastjsonschema==2.20.0
fastrlock==0.8.3
ffmpy==0.3.2
filelock==3.17.0
fire==0.7.0
flake8==7.1.2
flash_attn==2.7.2.post1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
freetype-py==2.4.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.9.0
ftfy==6.3.1
gdown==5.2.0
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.66.0
gradio==4.39.0
gradio_client==1.1.1
grpcio==1.65.1
h11==0.14.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.28.1
humanfriendly==10.0
icecream==2.1.4
idna==3.10
imageio==2.34.2
imageio-ffmpeg==0.5.1
immutables==0.20
importlib_metadata==8.1.0
importlib_resources==6.4.0
inflect==7.3.1
iniconfig==2.0.0
inquirerpy==0.3.4
interegular==0.3.3
ipdb==0.13.13
ipython==8.26.0
ipywidgets==8.1.3
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.11
kagglehub==0.3.11
kapture==1.1.10
kapture-localization==1.1.10
kiwisolver==1.4.5
kornia==0.5.10
lark==1.2.2
latex2sympy2_extended==1.10.1
lazy_loader==0.4
librosa==0.10.2.post1
liger_kernel==0.5.3
lightning==2.4.0
lightning-utilities==0.11.6
linkify-it-py==2.0.3
llm-blender==0.0.2
llvmlite==0.43.0
lm-format-enforcer==0.10.9
lmdeploy==0.5.3
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
math-verify==0.7.0
matplotlib==3.9.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.16.0
mergekit==0.0.6
mistral_common==1.5.3
mmengine-lite==0.10.5
modelscope==1.23.1
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
msgspec==0.19.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbformat==5.10.4
nemo_text_processing==1.0.2
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.3
numba==0.60.0
numpy==1.26.4
numpy-quaternion==2023.0.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
onnxruntime==1.18.1
open3d==0.18.0
openai==1.57.4
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
optimum-quanto==0.2.6
orjson==3.10.6
osqp==0.6.7.post1
oss2==2.19.1
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pairless==0.1.0
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
partial-json-parser==0.2.1.1.post5
pathspec==0.12.1
peft==0.10.0
pexpect==4.9.0
pfzy==0.3.4
piexif==1.1.3
pillow==10.4.0
pillow_heif==0.17.0
platformdirs==4.2.2
plotly==5.23.0
pluggy==1.5.0
plyfile==0.8.1
pooch==1.8.2
poselib==2.0.2
proces==0.1.7
proglog==0.1.10
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.47
proto-plus==1.26.0
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase16384==0.3.7
pybind11==2.13.6
pycocoevalcap==1.2
pycocotools==2.0.8
pycodestyle==2.12.1
pycolmap==3.10.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
pyglet==1.5.29
Pygments==2.18.0
pyheif==0.8.0
PyMatting==1.1.12
pynini==2.1.5
pynvml==12.0.0
PyOpenGL==3.1.0
pyparsing==3.1.2
pyproject==1.3.1
pyproject_hooks==1.1.0
pyquaternion==0.9.9
pyrender==0.1.45
PySocks==1.7.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-rerunfailures==15.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytorch-ranger==0.1.1
pytorch-triton==3.0.0+dedb7bdf33
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.2
pyzmq==26.2.1
qdldl==0.1.7.post4
qwen-vl-utils==0.0.8
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
rembg==2.0.58
requests==2.32.3
retrying==1.3.4
rich==13.7.1
roma==1.5.0
rpds-py==0.19.0
rsa==4.9
ruff==0.5.4
s5cmd==0.2.0
sacremoses==0.1.1
safetensors==0.5.2
scikit-image==0.21.0
scikit-learn==1.5.1
scipy==1.14.0
scs==3.2.6
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
shellingham==1.5.4
shortuuid==1.0.13
simple_knn==0.0.0
six==1.16.0
sk-video==1.1.10
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
textual==2.1.2
threadpoolctl==3.5.0
tifffile==2024.7.24
tiktoken==0.8.0
timm==0.9.12
tinycudann==1.7
tokenizers==0.21.0
tomli==2.0.1
tomlkit==0.12.0
torch==2.5.1
torch-optimizer==0.3.0
torchaudio==2.5.1
torchmetrics==1.4.0.post0
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trimesh==4.4.3
triton==3.1.0
typeguard==4.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.30.3
uvloop==0.19.0
vector-quantize-pytorch==1.15.3
virtualenv==20.29.2
vllm==0.7.3
vocos==0.1.0
wandb==0.19.8
watchfiles==0.22.0
Wave==0.0.2
wcmatch==8.5.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.3
WeTextProcessing==1.0.3
wget==3.2
whatimage==0.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
xformers==0.0.28.post3
xfuser==0.4.1
xgrammar==0.1.11
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.9.4
yunchang==0.6.0
zhconv==1.4.3
zhon==2.0.2
zipp==3.19.2
